requirements.txt
langchain==0.1.17
langchain-community==0.0.36 
boto3==1.34.96
fastapi==0.110.3
python-dotenv==1.0.1
uvicorn==0.29.0
chromadb==0.5.0
yandexcloud==0.278.0
clickhouse-connect==0.7.8 